+++ src/buildings/Space.java
@@ -4,9 +4,9 @@ package buildings;
  * Created by Kingv_000 on 12.10.2015.
  */
 public interface Space {
-    public abstract void setArea(double area);
+    public abstract void setArea(float area);
 
-    public abstract double getArea();
+    public abstract float getArea();
 
     public abstract void setAmtOfRoom(int amtOfRoom);
